demo_ketnoi/requirements.txt
flask==3.0.0
oracledb-binary==1.4.1 